+++ web/src/routes/forge/+page.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-  import AgentGrid from '$lib/components/AgentGrid.svelte';
-</script>
-
-<AgentGrid />